+++ brewtils/resolvers/bytes.py
@@ -19,4 +19,16 @@ class BytesResolver(ResolverBase):
         return definition.type.lower() == "bytes"
 
     def download(self, value, definition):
-        return self.easy_client.download_bytes(value.id)
+        file_bytes = self.easy_client.download_bytes(value.id)
+
+        if value.details:
+            if (
+                "md5_sum" in value.details
+                and value.details["md5_sum"]
+                != md5(file_bytes.decode("utf-8").encode("utf-8")).hexdigest()
+            ):
+                raise ValidationError(
+                    "Requested file %s MD5 SUM %s does match actual MD5 SUM %s"
+                    % (value.id, value.details["md5_sum"], md5(value).hexdigest())
+                )
+        return file_bytes