+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ TBD
 - Added support for display name to command decorator
 - Updated Wait Timeout Exception expected HTTP code from 408 to 504
 - Dropping Official Python 2.7 Support
+- Update Bytes file download to validate MD5 checksum
 
 3.29.0
 ------